+++ src/components/login/LoginForm.tsx
@@ -2,12 +2,8 @@ import styled from '@emotion/styled';
 import LogoImg from '@/Assets/icons/logo.png';
 import LoginButton from '@/components/login/LoginButton';
 
-import axios from 'axios';
-import { useAuth } from '@/context/AuthContext';
 import { useLoginForm } from '@/hooks/useLoginForm';
-import { useNavigate, useLocation } from 'react-router-dom';
-import { login } from '@/api/loginApi';
-import { toast } from 'react-toastify';
+import { useLoginHandler } from '@/hooks/useLoginHandler';
 
 const Container = styled.div`
   width: 100%;
@@ -57,43 +53,12 @@ const LoginForm = () => {
     handlePasswordBlur,
   } = useLoginForm();
 
-  const navigate = useNavigate();
-  const location = useLocation();
-  const from = location.state?.from?.pathname || '/';
-  const { login: setUser } = useAuth();
+  const { handleLogin } = useLoginHandler();
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
     if (!isValid) return;
-
-    try {
-      const res = await login({ email, password });
-      const { authToken, email: userEmail } = res.data;
-
-      const extractedName = userEmail.split('@')[0];
-
-      const userInfo = {
-        email: userEmail,
-        name: extractedName,
-      };
-
-      setUser(userInfo);
-      localStorage.setItem(
-        'userInfo',
-        JSON.stringify({ authToken, email: userEmail, name: extractedName })
-      );
-
-      navigate(from, { replace: true });
-    } catch (error: unknown) {
-      if (axios.isAxiosError(error)) {
-        const message =
-          error.response?.data?.message ||
-          '로그인에 실패했습니다. 이메일 형식이 kakao.com으로 끝나는지 확인해주세요.';
-        toast.error(message);
-      } else {
-        toast.error('알 수 없는 오류가 발생했습니다.');
-      }
-    }
+    await handleLogin(email, password);
   };
 
   return (
