+++ src/hooks/useLoginHandler.ts
@@ -1,0 +1,41 @@
+import { useNavigate, useLocation } from 'react-router-dom';
+import { useAuth } from '@/context/AuthContext';
+import { login } from '@/api/loginApi';
+import axios from 'axios';
+import { toast } from 'react-toastify';
+
+export const useLoginHandler = () => {
+  const navigate = useNavigate();
+  const location = useLocation();
+  const from = location.state?.from?.pathname || '/';
+  const { login: setUser } = useAuth();
+
+  const handleLogin = async (email: string, password: string) => {
+    try {
+      const res = await login({ email, password });
+      const { authToken, email: userEmail } = res.data;
+
+      const extractedName = userEmail.split('@')[0];
+      const userInfo = { email: userEmail, name: extractedName };
+
+      setUser(userInfo);
+      localStorage.setItem(
+        'userInfo',
+        JSON.stringify({ authToken, email: userEmail, name: extractedName })
+      );
+
+      navigate(from, { replace: true });
+    } catch (error: unknown) {
+      if (axios.isAxiosError(error)) {
+        const message =
+          error.response?.data?.message ||
+          '로그인에 실패했습니다. 이메일 형식이 kakao.com으로 끝나는지 확인해주세요.';
+        toast.error(message);
+      } else {
+        toast.error('알 수 없는 오류가 발생했습니다.');
+      }
+    }
+  };
+
+  return { handleLogin };
+};
